api security, integretion JWT api

+++ user-api/delComment.php
@@ -1,6 +1,8 @@
 <?php
     require('../connectDB.php');
     require('../header.php');
+    require('../TokenManager.php');
+    require('../autoload.php');
 
     global $conn;
 
@@ -13,42 +15,48 @@
         return;
     }
 
-    if(isset($data->id_comment) && !empty($data->id_comment)){
-        $query = $conn->prepare('DELETE FROM HiChat.COMMENTAIRE 
-                                    WHERE
-                                    COMMENTAIRE.id_commentaire = :id_comment
-                                ');
-            $query->execute([
-                ':id_comment' => $data->id_comment
-            ]);
+    if(verifiedToken($_SERVER['HTTP_AUTHORIZATION'])){
+        if(isset($data->id_comment) && !empty($data->id_comment)){
+                $query = $conn->prepare('DELETE FROM HiChat.COMMENTAIRE 
+                                            WHERE
+                                            COMMENTAIRE.id_commentaire = :id_comment
+                                        ');
+                    $query->execute([
+                        ':id_comment' => $data->id_comment
+                    ]);
 
-            if($query){
-                    $response = [
-                    'success' => true,
-                    'valid' => true
-                    ];
-                echo json_encode($response);
+                    if($query){
+                            $response = [
+                            'success' => true,
+                            'valid' => true
+                            ];
+                        echo json_encode($response);
+                    }else{
+                        echo json_encode([]);
+                    }
             }else{
-                echo json_encode([]);
+                $query = $conn->prepare('DELETE FROM HiChat.COMMENTAIRE 
+                                            WHERE
+                                            COMMENTAIRE.PID = :id
+                                        ');
+                    $query->execute([
+                        ':id' => $data->PID,
+                    ]);
+
+                    if($query){
+                            $response = [
+                            'success' => true,
+                            'valid' => true
+                            ];
+                        echo json_encode($response);
+                    }else{
+                        echo json_encode([]);
+                    }
             }
     }else{
-        $query = $conn->prepare('DELETE FROM HiChat.COMMENTAIRE 
-                                    WHERE
-                                    COMMENTAIRE.PID = :id
-                                ');
-            $query->execute([
-                ':id' => $data->PID,
-            ]);
-
-            if($query){
-                    $response = [
-                    'success' => true,
-                    'valid' => true
-                    ];
-                echo json_encode($response);
-            }else{
-                echo json_encode([]);
-            }
+        http_response_code(403);
     }
+
+   
    
 ?>
