+++ user-api/deleteAccount.php
@@ -1,6 +1,8 @@
 <?php
     require('../connectDB.php');
     require('../header.php');
+    require('../TokenManager.php');
+    require('../autoload.php');
 
     global $conn;
 
@@ -12,22 +14,29 @@
     }else{
         return;
     }
-        $query = $conn->prepare('DELETE FROM HiChat.USERS 
-                                    WHERE
-                                    USERS.id_USERS = :idUser
-                                ');
-            $query->execute([
-                ':idUser' => $data->id_users
-            ]);
 
-            if($query){
-                    $response = [
-                    'success' => true,
-                    'valid' => true
-                    ];
-                echo json_encode($response);
-            }else{
-                echo json_encode([]);
-            }
-    
+    if(verifiedToken($_SERVER['HTTP_AUTHORIZATION'])){
+            $query = $conn->prepare('DELETE FROM HiChat.USERS 
+                                                WHERE
+                                                USERS.id_USERS = :idUser
+                                            ');
+                        $query->execute([
+                            ':idUser' => $data->id_users
+                        ]);
+
+                        if($query){
+                                $response = [
+                                'success' => true,
+                                'valid' => true
+                                ];
+                            echo json_encode($response);
+                        }else{
+                            echo json_encode([]);
+                        }
+                
+    }else{
+        http_response_code(403);
+    }
+
+        
 ?>
